rename modifiers package
